+++ packages/api/src/decorator/ActivityGrouping/types.ts
@@ -1,1 +1,0 @@
-export const activityGroupingDecoratorTypeName = 'activity grouping' as const;