+++ packages/api/src/hooks/internal/private/templateMiddleware.ts
@@ -1,4 +1,0 @@
-import templateMiddleware from '../../../decorator/private/templateMiddleware';
-
-// TODO: We should move them to a common directory.
-export default templateMiddleware;